+++ src/assembler/utilities.h
@@ -12,6 +12,6 @@
 #include "asm.h"
 
 extern isa_register_t convertBinaryStringToInt(char *binaryString, uint64_t numberOfBitsInString);
-extern char* convertValueToBinaryString(int value, size_t length);
+extern char* convertValueToBinaryString(uint64_t value, size_t length);
 
 #endif
